Login + CRUD
Login page with ability to create, update and delete questions.

Note: This is an MVP login system - it's just designed to store API Keys which it can send with Create, Update and Delete requests to needs of admin users for now (there is no server side session tracking; it just saves the users API locally and uses that for subsequent requests).

Cookie based token check in works with server side rendering, but not for actual login or logout (which still requires client side JavaScript). This puts us in a good position for Next.js can handle custom routes for POST requests - when that is in place we will be able to have 100% universal rendering, but for now requiring JavaScript for site admins to login is fine.