Fixing a bug triggered when question view updates
Keeping the question in this.state (and updating it later when related questions had been populated via an async all) occationally triggered a bug.

This was possibly made worse with the introduction of the new markdown parser, but it’s hard to debug and confirm the issue.

The question is now only set as a prop and just relatedQuestions  is stored in the state, and safely updates.