Updating config for deployment on AWS EB
Adding a .ebignore file to control what files Elastic Beanstalk will include when deploying via CLI.

By default it uses .gitignore, which in this case includes the .next and node_modules directories, but we want these to be the same for each install.

This especially true for the pre-build assets in .next as otherwise, due to the way webpacked assets are build with hashes in filenames, it will break how load balancing works with the application.