Updating .ebingore to attempt to resolve bug
Deployments are failing, possibly due to node_modules (which I can see would potentially cause a problem and is a bad idea as deployments are coming from a different OS).

Alternatively, .ebignore is not working as suggested and I will need to handle it differently by generating a zip file with all the assets I want to include.