Issue with EBS debugged, but not resolved yet
The problem with the deploys to Elastic Beanstalk are:

1. If you let each server build it’s own .next assets dir then the hashes will be different and users will get errors loading pages (as they will be servered from different hosts.

2. If you deploy a prebuilt .next assets folder, it hardcodes in the path of some modules (specifically babel, it seems).

So, unless I can persuade the Next.js team to patch this by perhaps generating hash that is predeterminable and so consistant on all hosts (e.g. based on a file hash) I would need to build it on one target machine then clone deploy that same image to all instances in the autoscaling group, and I’m not sure there is an easy way to do that.

As a short term measure, I can enable sticky load balancing. This is not ideal as users may sometimes still experience errors and it makes managing load more difficult.

I’m glad I know what the problem is, but it’s quite a show stopper as it’s a pain to work around with EBS as it would require quite a bit of scripting to set something up to work around this.