Updating to Beta 18 and new build/deploy scripts
The build process now hooks into a feature in Beta 18 where the BUILD_ID can be customised after building is complete.

The BUILD_ID is now tied to the application version (as defined in package.json) so it’s incrimental, unique to each build and predictable.

This is helpful when deploying to the EBL cluster as it means the BUILD_ID “hash” will be consistant across all servers (instead of each server generating it’s own ID, which causes problems fetching resoruces unless sticky sessions are enabled).