Refactoring to reduce warnings/errors
Have updated now <Links> are rendered to remove warnings (as the pragma has changed in Next.js).

QuestionCards are now a component that is reused on the index, search and question pages, reducing code duplication.

Minor tweak to visual style of question cards.

Have temporarily removed videos and images when redering cards or questions, until the feature is ready for production.