Refactoring to New, Edit and Menubar components
The New Question and Edit Question views now share a common compontent, helping to ensure consistant rendering and behaviour in both views.

The Menubar now renders at the top of the view when logged in. This avoids distrupting the flow of how the page is displayed when logged in, giving a more consistent appearance whether signed in or not.