Move sass generation into the app
This removes the exernal dependancy on the ‘sass’ command line app. I would like to find a better way of hooking into the webpack process to generate a CSS file in the _next dir during (or after) the next build process. I am hoping this is something that makes it natively into next 2.0, but may refactor what I have done into a library if not.