Moved window.scrollTo(0,0) to componentWillUnmount
This casues the window to scroll to the top just before loading a new page (instead of as soon as the new page is initalised).

This means it loads as soon as the new page is ready and not before (which is confusing behaviour, as what was happening in practice is it was scrolling to the top of the current page before navigating away).