Adding menubar with Home and Subscribe links
Also changed rendering for cards used on the home and search pages, as CSS animiation approach being used caused most previewers to render th emain conent area of the homepage blank.

Not sure of a way to achive the same loading effect without causing that issue. Have left it in place on individual question pages.