Forced to had in hack to GA tracker :(
As there is currently no easy way I can see to only update once route is loaded (although it is possible, probably with _document.js) I’ve resorted to just adding in a 500ms delay in the updateGoogleAnalytics call which is ugly and impacts time on page tracking metrics but fixes the issue of live tracking showing the wrong medatadata.

We can come up with a better fix in future.