Removing react-reponsive as buggy
When going from large window to small window doesn’t render properly (but does if you *start* from a small window, make it bigger, then shrink it).

This bug was casued by the MediaQuery having the values={{width: 660}} set, but if you remove that then it doesn’t work at all in MSIE (any version).

Could delve in and fix it and submit a pull request but the whole thing is simple to fix (and avoids pulling in a lib and rolling in 3 unique dependancies!) by using CSS directives so just doing that.