Force redirect to https://upsum.news base URL
All requests to //www.upsum.news should now redirect to the https://upsum.news/ version of the URL (maintaining the path and any query string).

This is to ensure consistancy of URL as it appears in search engine results. HTTP to HTTPS redirection is already handled automatically by our cloud platform.