Adding hook to update front page without reload
The front page will now check for recent items in the background on page load (it will still load immediately when going back to the homepage but will update instantly if there are any new stories since the reader last looked).

Also includes updates to feed and sitemap docs and cruft removal.