Now caches trending questions
Caches trending questions on homepage and sidebar.

The cache currently updates every 5 minutes, although items *may* appear in RSS and Facebook Instant article feeds sooner.

It caches them in memory to reduce the overhead of expensive queries to the API, especially as approach being used (a full text search) wasn’t very fast; also the nature of the current cloud setup means a request goes through multiple servers, potentially in different regions and zones, so it’s a little slow.

This approach is a bit of a hack until we revise the backend API.