Adding dedicated RSS feed for email with images
This feed adds an image to every 3 articles (including the first one), to add some images to the automated email but too many (so it’s not overwhelming).

This has been done along side some improvements to the format of the email in MailChimp. This is still not as good as having a human curated email (or even bot curated one - e.g. based on views/trending terms) but is a significant improvement in the attractiveness of the daily emails.