Updating AdSense behaviour
It looks like we’ll need to switch to using AdSense with DoubleClick for Publishers (DFP) to use AdSense with our Single Page App.

I’ve tried a spike with DFP and it’s going to take a bit of work to get it into shape and to a point where it will work as a drop in replacement (in particular if we can get it to work well with responsive slots, which would be ideal).

For now, I’ve settled on a compromise of updating the existing AdSense code to improve the experience to the point where it at least shouldn’t be buggy.

We can do the work to integrate with DFP when we are at the point of having a commercial case to justify doing it.