Improving / masking potentially fatal errors
There are still errors about not being able to read the “.status” property which then cause the app to stop responding to future link clicks.

It may be an internal error with Next.js or it could be a problem triggered by requests failing when there is client/sever connection issue;  going to try refactoring and wrapping the only couple of things that could be triggering the problem to see if it helps.